Fix for second temporary token being interpreted as success
The code now keeps track of whether or not the final tokens were
actually requested.

Before if we received temporary tokens a second time it would behave as
if the authorization succeeded and would simply fail any actual
authorizing and would stop attempts to authorize again. By keeping
track of whether or not an actual request was made we can now tell the
temporary and the final tokens apart.

Change-Id: I4a55e7beda9e7be9d8c693739f4d3985438837ea
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Jesus Fernandez <Jesus.Fernandez@qt.io>