Tests now check for host's availability
oauthbin.com was down (domain name expired) between June 23rd to
July 6th with no information on whether or not it would come back.

Because of this the testdata is being changed to skip unavailable
resources.

These tests are already bypassed in CI and will not affect any of the
tests run in CI. For CI a more stable solution should be created later.

Change-Id: I50e302153a7b8be823e7a4e97cee40350099f59e
Reviewed-by: Jesus Fernandez <Jesus.Fernandez@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>