Fix the twittertimeline example
The ‘twittertimeline’ example wasn’t working because it couldn’t find
a callback URL to redirect the user to. This was fixed by forcing the
HTTP server to listen to a specific port (1337), as it’s done in the
Reddit example. This commit also updates the documentation to explain
how the users should configure the callback URL on their Twitter app
settings.

Task-number: QTBUG-68632
Change-Id: Idaaa550ecc37b42f91efa80f63b03d9c222c2e96
Reviewed-by: Jesus Fernandez <Jesus.Fernandez@qt.io>