Update commercial license headers
Updated header.COMM to files in tqtc-qtnetworkauth.
Examples, documentation files or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4172
Change-Id: Ifb1913c903f76d6428beb49a2e3e2fc0f3648059
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>