wasm: Work around compilation failures when FEATURE_http=OFF
Some of the header and source files incorrectly check for QT_NO_HTTP
without including the necessary QtNetwork/qtnetwork-config.h
header. Because CMake's AUTOGEN still runs moc on those files and then
includes them into mocs_compilation.cpp, this causes compilation
failures.

Work around the compilation failures by including the moc files
directly in the respective cpp files.

Note that this ensures archiving works in a static build, but results
in a an unusable library. This was the case for Qt 5.15 as well.

A follow up change will be done to clean up this state, but for now
this change unblocks adding the WebAssembly platform into CI.

Task-number: QTBUG-78647
Task-number: QTBUG-93243
Change-Id: I9b6657d5ae539fdb09a741637cc5a19e652c9272
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>