CMake: Skip building QtNetworkAuth if QT_FEATURE_http is OFF
When configuring Qt with the http feature disabled (which happens when
the thread feature is disabled e.g. when targeting WebAssembly)
NetworkAuth compiled some files, but the static library was
unusable. In shared builds the library failed to link at all.

Skip the module if http feature is disabled.

This shouldn't break our CI test phase when the repo isn't built
because all such configs we have in the CI are static builds
(wasm) marked with DisableTests and thus no test phase is created.

Task-number: QTBUG-93243
Change-Id: I3831f4722c1dcf854cabe3d49cd81668dac0a0a4
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>