Conan: Inherit recipe class from QtLeafModule for common functionality
The qt-conan-common package implements a base class for Qt leaf
module recipes. The build steps in leaf modules are mostly identical
so it makes sense to put those in the base class.

Dependencies are read by the base class from the 'dependencies.yaml'
which is the same file the CI system uses.

Pick-to: 6.2
Task-number: QTBUG-94877
Change-Id: Id115544b59bd454ba43971df60d6d613ed6c8b0a
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Toni Saario <toni.saario@qt.io>