Update commercial license headers
Updated header.COMM to the files in tqtc-qtnetworkauth.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4952
Change-Id: I9bb1ad060b4cf33dbc4933b38788fd083b946fa0
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>