Ensure the code is not encoded a second time if already percent encoded
Google recently changed the way they send their code and it is now
already percent encoded. This patch checks for the percent presence in
the code and does not change it again if it's there.

The alternative is to use:
google = new QOAuth2AuthorizationCodeFlow;
// Setup authentication parameters
google->setModifyParametersFunction(
    [](QAbstractOAuth::Stage stage,
       QMultiMap<QString, QVariant>* parameters) {
  if (stage == QAbstractOAuth::Stage::RequestingAccessToken) {
    QByteArray code = parameters->take("code").toByteArray();
    parameters->insert("code", QUrl::fromPercentEncoding(code));
  }
});

[ChangeLog][OAuth2] OAuth2 providers might be sending the authentication
code already percent encoded. This is the case of Google. This now a
supported use case and the code is not systematically encoded anymore.

Fixes: QTBUG-81624
Change-Id: I43d66223a2aedf01fe0996de6798acc6d881c16b
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>