QAbstractOAuth: use QByteArray::toBase64() to encode random to text
The alphabet we had was 62 characters long, which made for clumsy
division/modulus operations. If we extend to 64, the operations are
simpler. And if we're going to do that anyway, we may as well use
QByteArray to encode pure random data.

That's what I use to generate random strings, actually:
 $ dd if=/dev/random bs=3 count=4 status=none | basenc --base64url
 IjlwaN9HD_7HeNX0

Pick-to: 6.5 6.7 6.8
Change-Id: Ie30a3caf09ef4176bb36fffd17cd9b2c7a5cf58c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>