Deprecate OAuth2 HTTP methods (get(), post(), put(), etc)
With OAuth2 these methods effectively add an
"Authorization: Bearer <token>" header.

This isn't adequate justification for providing a parallel and
inferior HTTP method API; it is better to guide users to use
QtNetwork APIs directly.

The 'inferior' here meaning that the API and supported data
types are more limited, their implementation puts constraints on
the Content-Type header (must be urlencoded, or a JSON object).
Also the implementation does things like injecting a custom
'User-Agent' header.

The OAuth HTTP method API makes historical sense: with OAuth1
the per-message signing is a more complex requirement, further
complicated by the fact that in some cases the message body-data
is part of the signature creation (with urlencoded mime-type).

But for OAuth2, which is the de-facto OAuth version these days,
it is better to guide users to more standard Qt networking APIs.

Note 1: There are no QT_DEPRECATED_SINCE #ifdefs.
This is because the QAbstractOAuth pure virtuals, some of
which the now-deprecated methods override/define, would not
compile.

Note 2: The reason the QAbstractOAuth pure virtuals aren't
deprecated is that OAuth1 uses them too.

Pick-to: 6.8
Task-number: QTBUG-124329
Change-Id: I76956997dd9819cde25f95c47c4196c4931f6e83
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>