Doc: Fix documentation warnings for QOAuth1 class
Commit 8e5e1012f9a9a9f01a12f631a3ed34570b6a4c5c replaced QPair<> with
std::pair<> in the header, but not in the .cpp file. While this
compiles OK as QPair is an alias of std::pair, QDoc got confused
because of the different argument types.

Fix the warnings by adding explicit \fn signatures.

Enforce zero documentation warnings when testing documentation builds in
CI.

Pick-to: 6.8
Change-Id: I5d527de1f748f752450f61edc6b0e4d784309e7e
Reviewed-by: Paul Wicking <paul.wicking@qt.io>