Don't clear QAbstractOAuth2::scope upon empty server response
The implementation set the 'scope' unconditionally to scopes
the authorization server returned in access token response.
The returned and requested scopes can differ, and a server might
also omit the 'scope' altogether, causing the Qt classes 'scope'
to be just cleared.

The scope behavior is specified by OAuth2 RFC 6749, chapter 5.1.
Returning differing and empty scopes is common and acceptable.
In case of empty scope, we can assume that the server granted
the scope that was requested.

Furthermore amend the 'scope' property documentation to reflect
these two roles it serves. In a later Qt version we should
introduce two new properties which clearly separate these roles.

[ChangeLog][QAbstractOAuth2][Important Behavior Changes]
If the authorization server returns an empty 'scope' response,
the requested scope is not cleared anymore. Instead, it is
assumed that the requested 'scope' was granted.

Pick-to: 6.7 6.5
Fixes: QTBUG-66415
Change-Id: I0685507c3ac7633fcf1d177958ba45ec6fd87bd2
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit d867b77770c09c9b6a970edcbd8041cff6b64e4a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>