Validate 'state' property when set
The RFC 6749 defines the allowed 'state' characters.
Using characters outside of this range can lead to
state mismatches.

I tested this against major publicly available vendors,
and the returned state varies between vendors, and as a
result, some of them will result in a state mismatch and a
failing OAuth 2 authorization.

Note: Even sticking to allowed characters doesn't entirely
guarantee compatibility with all authorization servers.
For instance starting the 'state' with multiple spaces
seems to cause differing responses.

[ChangeLog][QAbstractOAuth2][Important Behavior Changes]
The 'state' property is checked for validity and setting a
'state' that contains illegal characters will not be allowed.

Change-Id: I1c2586f14f453906034908b814be6417ebeffd76
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>