OAuth2 new authorization redirect handler function
This commit introduces a new function for supplying the redirect
URL received at the end of the authorization stage. The URL
contains the necessary parameters for validating and continuing
with rest of the OAuth2 flow.

The new function is intended for use cases where the redirect
URL is captured by some other means than by using the existing
handlers main functionality. One notable use case for this
is using an embedded user-agent (Qt WebEngine) instead of the
system browser.

Followup task will document the usage with Qt WebEngine.

Task-number: QTBUG-127839
Change-Id: I39f2efc8be1da545fa068a547d39d74ffddf1f1b
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Ivan Solovev (OOO) <ivan.solovev@qt.io>