Add errorOccurred and deprecate error signal
Using just error as a signal name makes things confusing as it sounds
rather like a getter name (e.g. lastError from QSqlQuery). It has been
decided quite some time ago to use errorOccurred as a replacement in
other placed in the Qt code base as this name makes it clearer that an
error happened and is more in line with the Qt standard that wants clear
naming.

[ChangeLog][QAbstractOAuth2] The error signal is now deprecated and
replaced by errorOccurred. This will make the code handling errors
clearer to write and read.

Change-Id: Ibd6a4056735f6e9e4bec32525a2576cf41579a08
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>