Change QOAuth2DeviceAuthorizationFlow::refreshAccessToken() return type
Initially the slot was returning bool, but the similar slot in
QOAuth2AuthorizationCodeFlow has a void return type.

This patch adjusts the return type of QOAuth2AuthorizationCodeFlow's
version to void as well. That is done for consistency, and with a
potential future refactoring in mind.

The errors that can occur during the execution of the slot, are
already reported by the requestFailed() signal, so adjust the tests
to check this signal instead of the return value.

Amends 164e2d897fb7c3a60b518985774a4faa360ba2c9.

Change-Id: I0529ee354fe884b93d0ad9bb0f5eaa810ca7fcd1
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 551a63791b76aece906584379843b3c5eba8e8b1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>