Qt 7: add QAbstractOAuth2::refreshTokens() virtual slot
The base class already contains all properties and signals related to
access token refreshing. However, we can not add a new virtual slot
during Qt 6 lifetime.

This patch pre-programs the addition of this slot to Qt 7 and adjusts
the refresh token logic to call the new slot directly from the base
class. This eliminates the need to manually implement auto-refresh
logic in the derived classes.

The name of the new slot is different from the pre-existing slots in
the derived classes, so this patch also pre-programs the renaming
of the relevant slots.

The patch also pre-programs documentation changes by providing Qt 7
versions of the relevant docs.

Picking this change to 6.9 to minimize the amount of merge conflict
resolutions.

Task-number: QTBUG-132106
Change-Id: I453e9aae096abbfddcb8076f808d4a055850f7e0
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 2b1e3419c5ef43e964f40dbe79d7a7da0de5aa57)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>