Make tst_oauth2codeflow to use shared authorization server
In addition to fostering reuse and familiarity with
other test cases, this allows hiding the "HTTP details"
which are not the focus of these tests.

Note: commit modifies the test's reply handler to ingest
JSON objects instead of url-encoded data. This is just
an autotest detail since the test reply handler is not
a real handler.

Pick-to: 6.9
Task-number: QTBUG-131922
Change-Id: I604078544130b55ca6680f5b577a076dfda05eb7
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>