Document OAuth2 scope token character limitations
OAuth2 scope token characters are limited to a subset of printable
US-ASCII. This commit clarifies that characters outside of this
range are "not supported". The API (using QByteArray as the datatype)
and the implementation are optimized for RFC-compliant usage, while
issuing a helpful warning for violations of the RFC.

Public OAuth2 authorization servers appear to conform to the
RFC-defined character set. Use cases requiring characters outside
of the acceptable range are limited to the legacy 'scope' property,
as indicated by the runtime warning.

Pick-to: 6.9
Change-Id: I93a801b63e6375986027d6b3de8d7899b6467967
Reviewed-by: Marc Mutz <marc.mutz@qt.io>