Doc: Enable automatic linking to sources on code.qt.io
QDoc is capable of generating links to the declaration in the source
code for each documented C++ API entity.

Add the required configuration to enable this feature in Qt Network
Authorization. Specifically, the path to the root of the repository
was missing.

Fixes: QTBUG-147593
Change-Id: Ie77db91dd0f0eabceeefb666cc6d350623e72fb1
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>
(cherry picked from commit a13489346284df77dcf81ad1ea6bcda859445e00)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>